HypothesisUtils and Memoizer updated
HypothesisUtils.head_annotation added as a light weight way to check if
an annotation exists, a faster way to check for deleted annotations.

Memoizer.get_annos_from_api updated to explicitly sort by updated
and use ascending order to prevent gaps in case syncing fails.

Memoizer.h added so that a HypothesisUtils instance can be retrieved
more easily.

Memoizer.add_missing_annos updated to check the first block of
annotations to see if a subset of them are already present. This usually
means that annotations have been deleted from hypothesis. Also updated
to memoize annos a block at a time so that resync does not have to
complete all in one shot.