fix websocket not working issue
https://github.com/hypothesis/h/commit/c8f9f3e76af1e16cb7a66f3e1cf9782a31fcce83
seems to have been the cause of the websocket issues due to changes in
the schema which meant that include_all was no longer valid and
neither was [field]. Changing them to include_any and field solves the
issue.

An silent bug was also that /users should have been /user but since I
never used that filter the issue did not appear.