models group remove group scope related code
The stripped down database no longer has a notion of group scope at the
database level. Allow annotation creation anywhere, and filter after the
fact. If scopes are desired in the future then the current idea of scope
can be implemented as part of the basic workflow validation scheme. For
example a workflow assocated with a group can reject based on uri, tags,
text, or a combination of annotation type and any of the above, so that
replies could require one set of tags and page notes another set. If
someone wants to implement that logic in their custom client then they
can do that too, but memex doesn't need this in its core database.