HypothesisUtils added document field to payload
The Hypothes.is internal document info always sticks to the first
information about data document that it ever receives, so if there is no
title or other accompanying metadata about a document then it will
remain that way forever. This changes makes it possible to send document
data to the api. See recent scibot commits for an example of how to use
this to populate document data. In the long run the code from scibot
for document data extraction will probably move here since it is
critical for most web annotation applications.